+++ tests/test_basic_example.py
@@ -9,8 +9,6 @@
 # Note: These tests only function when the pytest output is not modified by plugins such as pytest-sugar!
 
 import utils
-import versions
-from packaging import version
 
 
 def test_deployment_failure(testdir):
@@ -32,8 +30,4 @@ def test_basic_example(testdir):
     utils.add_version_constraint_to_project(testdir.tmpdir)
 
     result = testdir.runpytest("tests/test_quickstart.py")
-
-    if versions.INMANTA_CORE_VERSION < version.Version("6"):
-        result.assert_outcomes(passed=2, skipped=1)
-    else:
-        result.assert_outcomes(passed=3)
+    result.assert_outcomes(passed=5)